chore: rename testing to fuzzing
